\name{compensationSet}
\alias{compensationSet}
\docType{data}
\title{ Sample Compensation Data Set}
\description{
  Sample Compensation Data Set
}
\usage{data(plateCore)}
\format{
  The format is an object of class \code{flowSet} composed of 5
  \code{flowFrames}.
  Each \code{flowFrame} corresponds to a compensation control, stained in
  one channel.
}

\details{

}
\source{
  Sample data set from BD FACS CAP analysis.
}

\keyword{datasets}
