\name{overlayFilterResult}
\alias{overlayFilterResult}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{ Result from an overlay filter operation }
\description{
  Contains the result from an \code{overlay} operation. \code{overlayFilterResults} are used
  for plotting and should not be accessed directly.
}
\usage{
overlayFilterResult(subSet, ...)
}
\arguments{
  \item{subSet}{ \code{numeric vector}}
}

\value{
  Returns an \code{overlayFilterResult}
}
\author{ Errol Strain }

\examples{
# \code{overlayFilterResults} should not be accessed directly.
}
% Add one or more standard keywords, see file 'KEYWORDS' in the
% R documentation directory.
\keyword{ methods }
