\name{pbmcPlate}
\alias{pbmcPlate}
\docType{data}
\title{ pbmcPlate Data Set}
\description{
  One plate from a BD FACS CAP analysis of Peripheral Blood Mononucleocyte cells.
}
\usage{data(plateCore)}
\format{
  The format is an object of class \code{flowSet} composed of 96
  \code{flowFrames}.
  Each \code{flowFrame} corresponds to one well from the plate.

}
\details{

}
\source{
  Complete dataset available at
  \url{http://www.ficcs.org/software.html#Data_Files}, the Flow Informatics and Computational
  Cytometry Society website (FICCS)  
}
\references{

}

\keyword{datasets}
